+++ app/vs_gst_qmlglsink/gstrtpreceiver.cpp
@@ -68,52 +68,6 @@ GstRtpReceiver::~GstRtpReceiver()
 
 }
 
-enum class VideoCodec {
-  H264=0,
-  H265,
-  MJPEG
-};
-static VideoCodec conv_codec(const QOpenHDVideoHelper::VideoCodec codec){
-    if(codec==QOpenHDVideoHelper::VideoCodecH264)return VideoCodec::H264;
-    if(codec==QOpenHDVideoHelper::VideoCodecH265)return VideoCodec::H265;
-    return VideoCodec::MJPEG;
-}
-
-static std::string gst_create_rtp_caps(const VideoCodec& videoCodec){
-  std::stringstream ss;
-  if(videoCodec==VideoCodec::H264){
-    ss<<"caps=\"application/x-rtp, media=(string)video, encoding-name=(string)H264, payload=(int)96\"";
-  }else if(videoCodec==VideoCodec::H265){
-    ss<<"caps=\"application/x-rtp, media=(string)video, encoding-name=(string)H265\"";
-  }else{
-    ss<<"caps=\"application/x-rtp, media=(string)video, encoding-name=(string)mjpeg\"";
-  }
-  return ss.str();
-}
-static std::string create_rtp_depacketize_for_codec(const VideoCodec& codec){
-  if(codec==VideoCodec::H264)return "rtph264depay ! ";
-  if(codec==VideoCodec::H265)return "rtph265depay ! ";
-  if(codec==VideoCodec::MJPEG)return "rtpjpegdepay ! ";
-  assert(false);
-  return "";
-}
-static std::string create_parse_for_codec(const VideoCodec& codec){
-  // config-interval=-1 = makes 100% sure each keyframe has SPS and PPS
-  if(codec==VideoCodec::H264)return "h264parse config-interval=-1 ! ";
-  if(codec==VideoCodec::H265)return "h265parse config-interval=-1  ! ";
-  if(codec==VideoCodec::MJPEG)return "jpegparse ! ";
-  assert(false);
-  return "";
-}
-static std::string create_out_h264(){
-    std::stringstream ss;
-    ss<<"video/x-h264";
-    ss<<", stream-format=\"byte-stream\"";
-    //ss<<", alignment=\"nal\"";
-    ss<<" ! ";
-    return ss.str();
-}
-
 static std::shared_ptr<std::vector<uint8_t>> gst_copy_buffer(GstBuffer* buffer){
   assert(buffer);
   const auto buff_size = gst_buffer_get_size(buffer);
@@ -155,12 +109,13 @@ static void loop_pull_appsink_samples(bool& keep_looping,GstElement *app_sink_el
 
 std::string GstRtpReceiver::construct_gstreamer_pipeline()
 {
-    const auto codec=conv_codec(m_video_codec);
+    const auto codec=pipeline::conv_codec(m_video_codec);
     std::stringstream ss;
-    ss<<"udpsrc address=\"127.0.0.1\" port="<<m_port<<" "<<gst_create_rtp_caps(codec)<<" ! ";
-    ss<<create_rtp_depacketize_for_codec(codec);
-    ss<<create_parse_for_codec(codec);
-    ss<<create_out_h264();
+    //ss<<"udpsrc address=\"127.0.0.1\" port="<<m_port<<" "<<gst_create_rtp_caps(codec)<<" ! ";
+    ss<<"udpsrc port="<<m_port<<" "<<gst_create_rtp_caps(codec)<<" ! ";
+    ss<<pipeline::create_rtp_depacketize_for_codec(codec);
+    ss<<pipeline::create_parse_for_codec(codec);
+    ss<<pipeline::create_out_h264();
     ss<<" appsink drop=true name=out_appsink";
     return ss.str();
 }
